Add type fix to other classes too

+++ khmer/_oxli/labeling.pyx
@@ -118,7 +118,7 @@ cdef class GraphLabels:
         return gl
 
     @staticmethod
-    def NodeGraphLabels(int k, int starting_size, int n_tables, primes=[]):
+    def NodeGraphLabels(int k, uint64_t starting_size, int n_tables, primes=[]):
         cdef vector[uint64_t] _primes
         if primes:
             _primes = primes
@@ -129,7 +129,7 @@ cdef class GraphLabels:
         return GraphLabels(graph)
 
     @staticmethod
-    def CountGraphLabels(int k, int starting_size, int n_tables, primes=[]):
+    def CountGraphLabels(int k, uint64_t starting_size, int n_tables, primes=[]):
         cdef vector[uint64_t] _primes
         if primes:
             _primes = primes
